Moving specialist mapping logic into specialist class.
