+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/repositories/jooq/JooqRepository.kt
@@ -1,12 +1,7 @@
 package org.vaccineimpact.api.app.repositories.jooq
 
 import org.jooq.DSLContext
-import org.vaccineimpact.api.serialization.Deserializer
-import org.vaccineimpact.api.serialization.UnknownEnumValue
-import org.vaccineimpact.api.app.errors.BadDatabaseConstant
 import org.vaccineimpact.api.app.repositories.Repository
+import org.vaccineimpact.api.app.repositories.jooq.mapping.MappingHelper
 
-abstract class JooqRepository(
-        val dsl: DSLContext,
-        val mapper: MappingHelpers = MappingHelpers()
-): Repository
+abstract class JooqRepository(val dsl: DSLContext): Repository
